+++ tests/files/root_siblings.xml
@@ -4,3 +4,4 @@
 <!-- a comment -->
 <?target some="processing" instructions="here"?>
 <root/>
+<!-- end -->
