--- tests/test_writing.py
@@ -1,73 +1,0 @@
-from delb import tag, Document, ParserOptions
-
-from tests.utils import assert_documents_are_semantical_equal, count_pis
-
-
-def test_initial_nodes(files_path, result_file):
-    Document(files_path / "initial_processing_instructions.xml").clone().save(
-        result_file
-    )
-    assert count_pis(result_file) == {
-        '<?another-target ["it", "could", "be", "anything"]?>': 1,
-        '<?target some="processing" instructions="here"?>': 2,
-    }
-
-    with result_file.open("rt") as file:
-        for line in file.readlines():
-            if "<!-- a comment -->" in line:
-                break
-        else:
-            raise AssertionError
-
-
-def test_significant_whitespace_is_saved(result_file):
-    document = Document("<text/>")
-    root = document.root
-    hi = tag("hi")
-
-    root.append_children(hi)
-    root[0].append_children("Hello")
-    root.append_children(" ")
-    root.append_children(hi)
-    root[2].append_children("world!")
-
-    document.save(result_file)
-    assert (
-        result_file.read_text() == "<?xml version='1.0' encoding='UTF-8'?>"
-        "<text><hi>Hello</hi> <hi>world!</hi></text>"
-    )
-
-    document.save(result_file, pretty=True)
-
-    assert (
-        Document(result_file, collapse_whitespace=True)
-        .xpath("hi")
-        .first.fetch_following_sibling()
-        == " "
-    )
-
-    with result_file.open("rt") as result:
-        assert result.readlines() == [
-            "<?xml version='1.0' encoding='UTF-8'?>\n",
-            "<text>\n",
-            "  <hi>Hello</hi>\n",
-            "   \n",  # FIXME?
-            "  <hi>world!</hi>\n",
-            "</text>\n",
-        ]
-
-
-def test_transparency(files_path, result_file):
-    for file in (x for x in files_path.glob("[!tei_]*.xml")):
-        doc = Document(file, parser_options=ParserOptions(collapse_whitespace=False))
-        doc.save(result_file)
-
-        assert_documents_are_semantical_equal(file, result_file)
-        assert count_pis(file) == count_pis(result_file)
-
-
-def test_xml_declaration(files_path, result_file):
-    Document(files_path / "tei_marx_manifestws_1848.TEI-P5.xml").save(result_file)
-    with result_file.open("rt") as f:
-        first_line = f.readline()
-    assert first_line.startswith("<?xml version='1.0' encoding='UTF-8'?>")
